Split bulk of state manipulation logic from GcodeHandlers into separate ExcludeRegionState class
